Fixed jre tests
